preparazione-esame/django/exercise_2-2_like/2-2_like4/pyproject.toml
[tool.poetry]
name = "exercise_2-2"
version = "0.1.0"
description = ""
authors = ["ilveron <alessandromonetti@outlook.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
django-cors-headers = "^4.3.1"
coreapi = "^2.3.3"
djangorestframework = "^3.14.0"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.55.0"
pytest-django = "^4.7.0"
mixer = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
